apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xox-api"
version = "0.1.0"
description = "Backend service for xox-model"
requires-python = ">=3.12"
dependencies = [
  "argon2-cffi>=25.1.0",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.38",
  "uvicorn>=0.35.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
